backend/requirements.txt
flask
flask-cors
requests
python-dotenv
cryptography
appwrite
sentence-transformers
scikit-learn
numpy
pandas
opencv-python
Pillow
imagehash
gunicorn